% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_inflection.R
\name{.as_param_vector}
\alias{.as_param_vector}
\title{Coerce a params vector/data frame to a named numeric vector a/b/c/d(/g)}
\usage{
.as_param_vector(params)
}
\description{
Coerce a params vector/data frame to a named numeric vector a/b/c/d(/g)
}
\keyword{internal}
