% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_limits.R
\name{.empty_detection_limits}
\alias{.empty_detection_limits}
\title{Empty detection_limits list for non-converged models}
\usage{
.empty_detection_limits(curve_id, model_name, method, alpha)
}
\description{
Empty detection_limits list for non-converged models
}
\keyword{internal}
