% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_limits.R
\name{.extract_param_ci}
\alias{.extract_param_ci}
\title{Extract parameter CIs from an ensemble entry}
\usage{
.extract_param_ci(ensemble_entry, method, alpha = 0.05)
}
\description{
Extract parameter CIs from an ensemble entry
}
\keyword{internal}
