% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_limits.R
\name{.interp_response}
\alias{.interp_response}
\title{Linear interpolation of predicted_response at a log10 concentration}
\usage{
.interp_response(x_grid, y_grid, x_target)
}
\description{
Linear interpolation of predicted_response at a log10 concentration
}
\keyword{internal}
