% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_limits.R
\name{.parabolic_refine}
\alias{.parabolic_refine}
\title{3-point parabolic interpolation for a set of extremum indices}
\usage{
.parabolic_refine(x, y, indices)
}
\description{
3-point parabolic interpolation for a set of extremum indices
}
\keyword{internal}
