% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_limits.R
\name{.safe_invert_model}
\alias{.safe_invert_model}
\title{Safely invert a named model at a response value}
\usage{
.safe_invert_model(model_name, y, params)
}
\description{
Safely invert a named model at a response value
}
\keyword{internal}
