% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{geom_mean}
\alias{geom_mean}
\title{Geometric mean}
\usage{
geom_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{Numeric vector (must be positive).}

\item{na.rm}{Logical. Remove NAs?}
}
\value{
Scalar geometric mean.
}
\description{
Geometric mean
}
\keyword{internal}
