% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradients.R
\name{grad_gompertz4}
\alias{grad_gompertz4}
\title{Analytical Gradient of the Inverse Gompertz}
\usage{
grad_gompertz4(y, a, b, c, d)
}
\arguments{
\item{y}{Numeric scalar. Observed response.}

\item{a, b, c, d}{Numeric scalars. Free model parameters.}
}
\value{
List with \code{grad_theta} and scalar \code{grad_y}.
}
\description{
Analytical Gradient of the Inverse Gompertz
}
\seealso{
Other gradient-functions: 
\code{\link{grad_logistic4}()},
\code{\link{grad_logistic5}()},
\code{\link{grad_loglogistic4}()},
\code{\link{grad_loglogistic5}()}
}
\concept{gradient-functions}
