% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradients.R
\name{grad_logistic4}
\alias{grad_logistic4}
\title{Analytical Gradient of the Inverse 4PL}
\usage{
grad_logistic4(y, a, b, c, d)
}
\arguments{
\item{y}{Numeric scalar. Observed response.}

\item{a, b, c, d}{Numeric scalars. Free model parameters.}
}
\value{
List with \code{grad_theta} (named vector) and scalar \code{grad_y}.
}
\description{
Returns \eqn{\partial x / \partial \theta} and \eqn{\partial x / \partial y}.
}
\seealso{
Other gradient-functions: 
\code{\link{grad_gompertz4}()},
\code{\link{grad_logistic5}()},
\code{\link{grad_loglogistic4}()},
\code{\link{grad_loglogistic5}()}
}
\concept{gradient-functions}
