% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradients.R
\name{grad_loglogistic5}
\alias{grad_loglogistic5}
\title{Analytical Gradient of the Inverse loglogistic5}
\usage{
grad_loglogistic5(y, a, b, c, d, g)
}
\arguments{
\item{y}{Numeric scalar. Observed response.}

\item{a, b, c, d}{Numeric scalars. Free model parameters.}

\item{g}{Numeric scalar. Asymmetry parameter.}
}
\value{
List with \code{grad_theta} and scalar \code{grad_y}.
}
\description{
Analytical Gradient of the Inverse loglogistic5
}
\seealso{
Other gradient-functions: 
\code{\link{grad_gompertz4}()},
\code{\link{grad_logistic4}()},
\code{\link{grad_logistic5}()},
\code{\link{grad_loglogistic4}()}
}
\concept{gradient-functions}
