% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverses.R
\name{inv_logistic5}
\alias{inv_logistic5}
\alias{inv_logistic5_fixed}
\title{Inverse of the 5PL Model}
\usage{
inv_logistic5(y, a, b, c, d, g, tol = 1e-06)

inv_logistic5_fixed(y, fixed_a, b, c, d, g)
}
\arguments{
\item{y}{Numeric vector. Observed response.}

\item{a, b, c, d}{Numeric scalars. Model parameters.}

\item{g}{Numeric scalar. Asymmetry parameter.}

\item{tol}{Numeric scalar. Buffer from asymptotes.}

\item{fixed_a}{Numeric scalar. Externally fixed lower asymptote.}
}
\value{
Numeric vector of \code{x} values. \code{NA} for out-of-range \code{y}.
}
\description{
Solves for \code{x}:
\deqn{x = c - b\,\log\!\left(\left(\frac{d - a}{y - a}\right)^{1/g}
  - 1\right)}
}
\seealso{
Other inverse-functions: 
\code{\link{inv_gompertz4}()},
\code{\link{inv_logistic4}()},
\code{\link{inv_loglogistic4}()},
\code{\link{inv_loglogistic5}()}
}
\concept{inverse-functions}
\keyword{internal}
