% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverses.R
\name{inv_loglogistic5}
\alias{inv_loglogistic5}
\alias{inv_loglogistic5_fixed}
\title{Inverse of the loglogistic5 Model}
\usage{
inv_loglogistic5(y, a, b, c, d, g)

inv_loglogistic5_fixed(y, fixed_a, b, c, d, g)
}
\arguments{
\item{y}{Numeric vector. Observed response.}

\item{a, b, c, d}{Numeric scalars. Model parameters.}

\item{g}{Numeric scalar. Asymmetry parameter.}

\item{fixed_a}{Numeric scalar. Externally fixed lower asymptote.}
}
\value{
Numeric vector of \code{x} values.
}
\description{
Solves for \code{x}:
\deqn{x = c - \frac{1}{b}\,\log\!\frac{
  \left(\frac{y - a}{d - a}\right)^{-g} - 1}{g}}
}
\seealso{
Other inverse-functions: 
\code{\link{inv_gompertz4}()},
\code{\link{inv_logistic4}()},
\code{\link{inv_logistic5}()},
\code{\link{inv_loglogistic4}()}
}
\concept{inverse-functions}
\keyword{internal}
