% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{resolve_effective_models}
\alias{resolve_effective_models}
\title{Resolve Effective Models for a Given Configuration}
\usage{
resolve_effective_models(fit_options, study_params)
}
\arguments{
\item{fit_options}{A \code{fit_options} object.}

\item{study_params}{A \code{study_params} object.}
}
\value{
Character vector of model names to actually fit.
}
\description{
Filters the requested model list based on whether the independent
variable is log-transformed (which makes loglogistic4 redundant
with logistic4).
}
